% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkfile.R
\name{checkfile}
\alias{checkfile}
\title{check whether files exist}
\usage{
checkfile(..., invert = FALSE)
}
\arguments{
\item{...}{files to check}

\item{invert}{check for existence of a file, instead of non-existence}
}
\value{
no return value
}
\description{
Check whether files exist (or don't exist), and stop()
}
\details{
I often want to run a script only if a particular file or files
already exist.  Or perhaps only run a script if the output file
doesn't already exist.  To avoid writing a lot of
=if(file.exists(f)) {}=
this is a convenience function.
}
\examples{
## Should do nothing
checkfile(system.file(c("INDEX","help")))
checkfile(system.file(c("ThisFileShouldNotExist")),invert=TRUE)
## Should call stop()
checkfile(system.file(c("INDEX","help")),invert=TRUE)
checkfile(system.file(c("ThisFileShouldNotExist")))
}
\author{
Chris Wallace
}
