% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cochran-Armitage.test.mod.R
\name{Cochran.Armitage.test}
\alias{Cochran.Armitage.test}
\title{Cochran Armitage test - with Mantel extension}
\usage{
Cochran.Armitage.test(exposure, cc, stratum = rep(1, length(cc)))
}
\arguments{
\item{exposure}{vector}

\item{cc}{vector}

\item{stratum}{(optional) vector of strata}
}
\description{
See Clayton & Hills 20.3 pp 201-203 for statistical details
}
\details{
This version is modified (bugfixed) from an original function by
David Clayton as below

change
x2 <- sum(ut^2/vt,na.rm=T)
to
x2 <- sum(ut,na.rm=T)^2/sum(vt,na.rm=T)

Modified Feb 2014 to change
ut <- dh[,1]*dh[,2]*(zm[,1] - zm[,2])/nt
to
ut <- dh[,1]*dh[,2]*(zm[,2] - zm[,1])/nt
so that sign(ut)>0 for positive correlation
}
\author{
David Clayton, with modifications by Chris Wallace
}
