% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtfun.R
\name{dt2mat}
\alias{dt2mat}
\title{data.table to matrix}
\usage{
dt2mat(dt, ...)
}
\arguments{
\item{dt}{data.table}

\item{...}{args passed to data.table::dcast}
}
\value{
a matrix
}
\description{
dcast a data.table, and return a matrix
}
\author{
Chris Wallace
}
