% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread-rownames.R
\name{freadrn}
\alias{freadrn}
\title{fread for files created with write.csv(..., row.names=TRUE)}
\usage{
freadrn(f)
}
\arguments{
\item{f}{filename}
}
\value{
fread object, ignoring first column (rownames)
}
\author{
Chris Wallace
}
