% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArgs.R
\name{getArgs}
\alias{getArgs}
\title{getArgs}
\usage{
getArgs(defaults = NULL, verbose = FALSE, numeric = NULL)
}
\arguments{
\item{defaults}{a named list of defaults, optional}

\item{verbose}{print verbage to screen}

\item{numeric}{names of arguments that should be converted from character to numeric, optional}
}
\value{
a named list
}
\description{
commandArgs parsing
}
\details{
return a named list of command line arguments

Usage:
call the R script thus
  ./myfile.R --args --myarg=something
or
  R CMD BATCH --args --myarg=something myfile.R

Then in R do
  myargs <- getArgs()
and myargs will be a named list

> str(myargs)
List of 2
$ file : chr "myfile.R"
$ myarg: chr "something"

To supply a vector as an argument, simply repeat the argument's name:

... --args f=kk f=pp ll

> str(myargs)
List of 2
$ f : Named chr [1:2] "kk" "pp"
 ..- attr(*, "names")= chr [1:2] "f" "f"
$ ll: logi TRUE
}
\author{
Chris Wallace
}
