% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-google-sheets.R
\name{getGSheets}
\alias{getGSheets}
\title{get google sheets as data.table}
\usage{
getGSheets(url, saveas = NULL)
}
\arguments{
\item{url}{google sheets sharing url}

\item{saveas}{if not null, save the csv format file at this location.  Must be a valid filename, and will get overwritten!}
}
\value{
data.table
}
\description{
get google sheets as data.table
}
\author{
Chris Wallace
}
