% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{ggheatmap.show}
\alias{ggheatmap.show}
\title{Display a ggheatmap}
\usage{
ggheatmap.show(L, col.width = 0.2, row.width = 0.2)
}
\arguments{
\item{L}{a list with 3 named plots: col, row, centre, generated by ggheatmap}

\item{col.width,row.width}{number between 0 and 1, fraction of the device devoted to the column or row-wise dendrogram. If 0, don't print the dendrogram}
}
\value{
no return value, side effect of displaying plot in current device
}
\description{
this function sets up some viewports, and tries to plot the dendrograms to line up with the heatmap
}
\author{
Chris Wallace
}

