% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{mysan}
\alias{mysan}
\title{Sanitizer function for xtable that converts scientific notation to latex style}
\usage{
mysan(x)
}
\arguments{
\item{x}{character or numeric vector}
}
\value{
sanitized character
}
\description{
Sanitizer function for xtable that converts scientific notation to latex style
}
\examples{
p<-10^(-seq(1,10))
format.pval(p) # scientific notation
mysan(p) # latex style

library(xtable)
df <- data.frame(numeric=p,scientific=format.pval(p),p.san=mysan(p))
print(xtable(df)) # doesn't work
print(xtable(df), sanitize.text.function=mysan)  # looks nice!
}

