% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowplots.R
\name{nicecow}
\alias{nicecow}
\title{nicify cowplot theme}
\usage{
nicecow(p)
}
\arguments{
\item{p}{plot object}
}
\value{
nicer plot object
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\examples{
library(ggplot2)
library(cowplot)
df <- data.frame(x=rnorm(1000),y=rnorm(1000))
p <- ggplot(df, aes(x=x,y=y)) + geom_point() + geom_density2d() + ggtitle("2D data")
p
nicecow(p)
}
\author{
Chris Wallace
}
