% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/random-functions-package.R
\docType{package}
\name{random-functions-package}
\alias{random-functions-package}
\title{title randomFunctions}
\description{
title randomFunctions
}
\author{
Chris Wallace <chris.wallace@cimr.cam.ac.uk>
}
\keyword{package}

