% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysave.R
\name{runorload}
\alias{runorload}
\title{run code or load previous run}
\usage{
runorload(file, objects, ...)
}
\arguments{
\item{file}{string, name of file to save in}

\item{objects}{character vector giving names of objects to save}

\item{...}{arguments to be evaluated. Use "<-" not "=" for any objects to be
saved!}
}
\value{
side effect: loads from file
}
\description{
load a file if it exists. Otherwise, run commands, and save created objects.
Experimental!
}
\examples{
runorload("tmp.RData","xnew",ls(),a=2,b=6,xnew<-a*b)
}
\author{
Chris Wallace
}
