% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textvenn.R
\name{textvenn}
\alias{textvenn}
\title{textvenn}
\usage{
textvenn(A, B, verbose = FALSE, quiet = FALSE)
}
\arguments{
\item{A}{vector}

\item{B}{vector}

\item{verbose}{if TRUE print all items to screen.  Default is FALSE.}

\item{quiet}{if FALSE (default) report summary counts to screen}
}
\value{
invisibly returns a list containing items unique to A, the
    intersection, and items unique to B
}
\description{
Find setdiff()s and intersection of two vectors
}
\author{
Chris Wallace
}
