\name{textvenn}
\alias{textvenn}
\title{textvenn}
\usage{
textvenn(A, B, quiet = FALSE)
}
\arguments{
  \item{A}{vector}

  \item{B}{vector}

  \item{quiet}{if FALSE (default) report to screen}
}
\value{
invisibly returns a list containing items unique to A, the
intersection, and intems unique to B
}
\description{
Find setdiff()s and intersection of two vectors
}
\author{
Chris Wallace
}

